osbuildexecutor: move tests to separate _test package
