+++ internal/osbuildexecutor/export_test.go
@@ -1,3 +1,8 @@
 package osbuildexecutor
 
+var ExtractOutputArchive = extractOutputArchive
+var FetchOutputArchive = fetchOutputArchive
+var HandleBuild = handleBuild
 var ValidateOutputArchive = validateOutputArchive
+var WaitForSI = waitForSI
+var WriteInputArchive = writeInputArchive
